More minor fixes to pave the way for fixing @LinkingObjects UTs
