+++ realm/realm-library/src/androidTest/java/io/realm/LinkingObjectsManagedTests.java
@@ -201,7 +201,8 @@ public class LinkingObjectsManagedTests {
                         assertEquals(2, looperThreadRealm.where(AllJavaTypes.class).findAll().size());
                     }
                 },
-                child, parent);
+                child,
+                parent);
     }
 
     // A listener registered on the backlinked field should be called when a commit adds a backlink
@@ -237,7 +238,8 @@ public class LinkingObjectsManagedTests {
                         assertEquals(1, counter.get());
                     }
                 },
-                child, parent);
+                child,
+                parent);
     }
 
     // A listener registered on the backlinked field should not be called after the listener is removed
@@ -273,7 +275,8 @@ public class LinkingObjectsManagedTests {
                         assertEquals(2, looperThreadRealm.where(AllJavaTypes.class).findAll().size());
                     }
                 },
-                child, parent);
+                child,
+                parent);
     }
 
     // A listener registered on the backlinked object should be called when a backlinked object is deleted
@@ -310,7 +313,8 @@ public class LinkingObjectsManagedTests {
                         assertEquals(1, counter.get());
                     }
                 },
-                child, parent);
+                child,
+                parent);
     }
 
     // A listener registered on the backlinked object should not called for an unrelated change
@@ -344,7 +348,8 @@ public class LinkingObjectsManagedTests {
                         assertEquals(1, looperThreadRealm.where(AllJavaTypes.class).findAll().size());
                     }
                 },
-                child, parent);
+                child,
+                parent);
     }
 
     // Fields annotated with @LinkingObjects should not be affected by JSON updates
@@ -450,6 +455,8 @@ public class LinkingObjectsManagedTests {
         final BacklinksTarget target = realm.where(BacklinksTarget.class)
                 .equalTo(BacklinksTarget.FIELD_ID, 1L).findFirst();
 
+        assertNotNull(target);
+
         realm.executeTransaction(new Realm.Transaction() {
             @Override
             public void execute(Realm realm) {
@@ -493,6 +500,8 @@ public class LinkingObjectsManagedTests {
             }
         });
 
+        assertNotNull(target);
+
         // precondition
         assertFalse(target.isValid());
 
@@ -503,14 +512,19 @@ public class LinkingObjectsManagedTests {
     }
 
     /**
-     * Table validation should fail if the backinked column already exists in the target table.
+     * Table validation should fail if the backlinked column already exists in the target table.
      * The realm `backlinks-fieldInUse.realm` contains the classes `BacklinksSource` and `BacklinksTarget`
-     * except that in the definition of `BacklinksTarget`, the field parent is defined as:
+     * except that in the definition of {@code BacklinksTarget}, the field parent is defined as:
      * <pre>
      * {@code
      *     private RealmList<BacklinksSource> parents;
      * }
      * </pre>
+     *
+     * <p/>
+     * The backlinked field does exist but it is list of links to {@code BacklinksSource} children
+     * not a list of backlinks to  {@code BacklinksSource} parents of which the {@code BacklinksTarget}
+     * is a child.
      */
     @Test
     public void migration_backlinkedFieldInUse() {
@@ -537,7 +551,7 @@ public class LinkingObjectsManagedTests {
     }
 
     /**
-     * Table validation should fail if the backinked column points to a non-existent class.
+     * Table validation should fail if the backlinked column points to a non-existent class.
      * The realm `backlinks-missingSourceClass.realm` contains two tables very like those
      * defined by `BacklinksSource` and `BacklinksTarget`.  In it, though, the source class
      * is named XXXBacklinksSource, like so:
@@ -547,10 +561,10 @@ public class LinkingObjectsManagedTests {
      *     private final RealmResults<XXXBacklinksSource> parents = null;
      * }
      * </pre>
-     * If the both classes were used in the configuration, the test would fail because of the
-     * missing class.  Since the configuration contains only the single class `BacklinksTarget`,
-     * basic validation passes.  Backlink validation, however, should fail, seeking the
-     * `BacklinksSource` table.
+     * If the both classes were present in the configuration, the test would fail because there
+     * is no class named BacklinksSource in the Realm.  Since the configuration contains only the
+     * single class `BacklinksTarget`, though, basic validation should pass. Backlink validation,
+     * however, should fail, seeking the `BacklinksSource` table.
      */
     @Ignore("Need to rebuild the test library")
     @Test
@@ -620,7 +634,7 @@ public class LinkingObjectsManagedTests {
     }
 
     /**
-     * Table validation should fail if the backinked column points to a field of the wrong type.
+     * Table validation should fail if the backlinked column points to a field of the wrong type.
      * This test is built in almost exactly the way as was `migration_backlinkedSourceFieldDoesntExist`
      * The realm `backlinks-sourceFieldWrongType.realm` was constructed with classes `BacklinksWrongTypeTarget`
      * and `BacklinksWrongTypeSource`.  Again, these two classes are nearly identical in their counterparts
