+++ realm/realm-library/src/main/java/io/realm/RealmConfiguration.java
@@ -728,7 +728,8 @@ public class RealmConfiguration {
          * create a module. These classes must be available in the default module. Calling this will remove any
          * previously configured modules.
          */
-        Builder schema(Class<? extends RealmModel> firstClass, Class<? extends RealmModel>... additionalClasses) {
+        @SafeVarargs
+        final Builder schema(Class<? extends RealmModel> firstClass, Class<? extends RealmModel>... additionalClasses) {
             if (firstClass == null) {
                 throw new IllegalArgumentException("A non-null class must be provided");
             }
